de-monolith Stage 6b: extract the shared kirra-map crate from the planner
Lift the planner's lane-map logic into a standalone, lean `kirra-map` crate — the
shared lane map originally requested, now unblocked by Stage 6a (which put the
corridor seam + perception data types in `kirra-core`). `kirra-map` depends ONLY on
`kirra-core` + `roxmltree`: no planner, no ROS2 adapter, no verifier service.

Moved verbatim into `kirra-map` (bodies proven code-identical by diff; only the
imports and a few cross-crate doc links changed):
- `kirra_map::lanemap` <- the planner's `lanemap.rs` (the Lanelet2-lite `LaneGraph` /
  `Lane` / `LaneCorridor` / `JunctionContext`, the router, right-of-way derivation,
  `from_undivided_corridor`). Its adapter imports are repointed to the lean source:
  `kirra_ros2_adapter::corridor::{CorridorSource,Point}` -> `kirra_core::corridor::*`;
  `::state::PerceivedObject` -> `kirra_core::trajectory::PerceivedObject`.
- `kirra_map::lanelet2` <- the planner's `lanelet2.rs` (the pure-Rust Lanelet2 `.osm`
  OSM-XML parser → `LaneGraph`). Carries the `roxmltree` dep with it.
- `kirra_map::lane_lines` <- the lane-line crossing rules (`LineType` / `LaneBoundary` /
  `may_cross` / `lateral_move_permitted`, + their 5 tests), split out of the planner's
  `behavior.rs`. The longitudinal traffic-control logic (`SignalState` / `TrafficControl`
  / `BehaviorConfig` / `Behavioral` / `evaluate_controls`) stays in the planner — it is
  policy, not map substrate.

The planner's `lanemap.rs` and `lanelet2.rs` become thin `pub use kirra_map::…::*;`
re-export shims, and `behavior.rs` re-exports the lane-line types from
`kirra_map::lane_lines`. So every existing path holds unchanged:
`kirra_planner::{LaneGraph, Lane, LaneCorridor, LaneEdge, JunctionContext, MAX_ROUTE_LANES,
LaneBoundary, LineType, parse_lanelet2_osm, Lanelet2ParseError}` and the internal
`crate::lanemap::*` / `crate::behavior::{LaneBoundary, LineType, lateral_move_permitted}` /
`crate::lanelet2::*` uses across `lib.rs` and `mick.rs`. The planner adds the `kirra-map`
dep and drops its now-unused direct `roxmltree` dep (it rides on `kirra-map`).

Verification: root workspace build + tests green — the 35 lane-map/lanelet2/lane-line
tests relocated cleanly from the planner (lib 102 -> 67) into `kirra-map` (35), none
lost; all planner integration suites (overtake_oncoming, unmarked_road_keep_right,
lane_graph_to_occy, taj_occy_kirra_stack, adversarial/learned doer) pass via the
re-exports. CI-equivalent `cargo clippy --workspace -- -D warnings` clean;
`kirra-map` rustdoc builds with `-D rustdoc::broken_intra_doc_links`; parko workspace
logic green (parko-core 204, parko-kirra 148, parko-ros2 168); no `#[path]` hazards;
both relocated modules proven code-identical by diff.
